u-boot-update: use U_BOOT_FDT as absolute path if it is one (Closes: #1025435)
u-boot-update would try to load U_BOOT_FDT from U_BOOT_FTD_DIR
in all cases, which is a hindrance to specifying an absolute
unversioned path for a device tree file to be loaded from.

Change this so that the first thing the script tries with regards
to generating the FDT line is to check whether U_BOOT_FDT begins
with a slash. If it does, and the file exists as an absolute path,
use it as the fdt file.