u-boot-update: dtb version by sed slicing & make fdtoverlays empty
If there are no overlays which are enabled then make FDTOVERLAYS variable
empty so that previous FDTOVERLAYS will not be written for different
kernel.

Setting set -e and doing grep was causing issue as the same string
is not found then process quits from that point which is not what
expected.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>